fix(components): Address code issues and improvements

+++ src/components/post-form/PostForm.jsx
@@ -4,6 +4,7 @@ import { Button, Input, Select, RTE } from "../index";
 import appwriteService  from "../../appwrite/config";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { toast } from "react-toastify";
 
 const PostForm = ({ post }) => {
     
@@ -20,37 +21,43 @@ const PostForm = ({ post }) => {
     const userData = useSelector(state => state.auth.userData);
 
     const submit = async (data) => {
-        if (post) {
-            const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+        try {
+            if (post) {
+                const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
 
-            if (file) {
-                appwriteService.deleteFile(post.featuredImage);
-            }
-            const dbPost = await appwriteService.updatePost(
-                post.$id, {
-                    ...data,
-                    featuredImage: file ? file.$id : undefined,
+                if (file && post.featuredImage) {
+                    await appwriteService.deleteFile(post.featuredImage);
                 }
+
+                const dbPost = await appwriteService.updatePost(
+                    post.$id, {
+                        ...data,
+                        featuredImage: file ? file.$id : undefined,
+                    }
                 )
-            if (dbPost) {
-                navigate(`/post/${dbPost.$id}`);
-            }
 
-        } else {
-            const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
-
-            
-            if (file) {
-                const fileId = file.$id;
-                data.featuredImage = fileId;
-                const dbPost = await appwriteService.createPost({
-                    ...data,
-                    userId: userData.$id
-                })
                 if (dbPost) {
+                    toast.success("Post updated successfully");
                     navigate(`/post/${dbPost.$id}`);
-                }            
+                }
+            } else {
+                const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+                
+
+                if (file) {
+                    data.featuredImage = file.$id;
+                    data.userId = userData?.$id || "";
+                    const dbPost = await appwriteService.createPost(data);
+    
+                    if (dbPost) {
+                        toast.success("Post created successfully");
+                        navigate(`/post/${dbPost.$id}`);
+                    }
+                }
             }
+        } catch (error) {
+            console.log("PostForm :: submit :: error", error);
+            toast.error(error.message);
         }
     }
